Updated the .gitignore
Not really a real update, but I saw that a file that likely isn't nessesary to be pushed, so I put it in the .gitignore. I also encourage everyone else to do the same when they see that they are pushing similar files.  Aka files that look to be more about Visual Studio than the actual project.